+++ server/app/Http/Controllers/ProjectController.php
@@ -41,13 +41,12 @@ class ProjectController extends Controller
 
 	public function add(Request $request) {
 		try {
+			$user = auth()->user();
 			$validatedData = $request->validate([
 				'name' => ['required', 'string', 'min:4', 'max:20', 'unique:projects'],
 				'description' => ['required', 'string', 'min:4', 'max:100'],
         'deadline' => ['required'],
         'payment_type_id' => ['required', 'string'],
-        'company_id' => ['required', 'string'],
-        'user_id' => ['required', 'string']
 			]);
 
 			$project = Project::create([
@@ -55,13 +54,14 @@ class ProjectController extends Controller
 				'description' => $validatedData['description'],
         'deadline' => $validatedData['deadline'],
         'payment_type_id' => $validatedData['payment_type_id'],
-        'company_id' => $validatedData['company_id'],
-        'user_id' => $validatedData['user_id'],
+        'company_id' => $user['company_id'],
+        'user_id' => $user['id'],
 			]);
 			return response()->json([
 				'message' => 'Проект успешно добавлен'
 			]);
 		} catch (\Throwable $th) {
+			echo $th;
 			return response()->json([
 				'message' => 'Не удалось добавить проект'
 			], 400);
@@ -70,10 +70,16 @@ class ProjectController extends Controller
 
 	public function update(Request $request, $projectId) {
 		try {
-			$project = PaymentType::where('id', $projectId)->first();
+			$project = Project::where('id', $projectId)->first();
 			if (!$project) {
 				return response()->json([
 					'message' => 'Проекта с таким ID не существует'
+				], 404);
+			}
+			// Дополнить или переделать
+			if (auth()->user()->id !== $project->user_id || auth()->user()->role->value !== 'ADMIN') {
+				return response()->json([
+					'message' => 'У вас нет прав на обновление данных проекта'
 				], 400);
 			}
 
@@ -82,8 +88,8 @@ class ProjectController extends Controller
 				'description' => ['required', 'string', 'min:4', 'max:100'],
         'deadline' => ['required'],
         'payment_type_id' => ['required', 'string'],
-        'company_id' => ['required', 'string'],
-        'user_id' => ['required', 'string']
+        'company_id' => $project['company_id'],
+        'user_id' => $project['user_id'],
 			]);
 
 			$project->update($validatedData);
@@ -91,6 +97,7 @@ class ProjectController extends Controller
 				'message' => 'Данные успешно обновлены'
 			]);
 		} catch (\Throwable $th) {
+			echo $th;
 			return response()->json([
 				'message' => 'Не удалось обновить данные проекта'
 			], 400);
@@ -103,6 +110,12 @@ class ProjectController extends Controller
 			if (!$project) {
 				return response()->json([
 					'message' => 'Проекта с таким ID не существует'
+				], 404);
+			}
+			// Дополнить или переделать
+			if (auth()->user()->id !== $project->user_id || auth()->user()->role->value !== 'ADMIN') {
+				return response()->json([
+					'message' => 'У вас нет прав на удаление проекта'
 				], 400);
 			}
 			$project->delete();
